Changed the name of several classes.
Changed the '*Listener' names to '*Consumer' names to better reflect their purpose.
